+++ docs/release-notes.md
@@ -16,6 +16,7 @@ Minor changes:
 - install: Verify that Ignition config is valid JSON
 - install: Add long help for several options
 - download/install: Clarify help text for `--insecure`
+- install: Run 'chreipl' only on s390x zVM
 
 Internal changes: